Listing 4-6: Another Horrible Mistake
Goofing up on purpose again
The section “Goofing up on purpose,” earlier in this chapter, introduces you
to compiler errors. The complier is only part of the program creation pro-
cess. The other major part is linking, and, yes, you’ll find that the linker can
detect errors as well, as shown in Listing 4-6.