Listing 5-1: Displaying Various Values
The output from Exercise 5-1 looks something like this:
The value 869 is an integer.
The value 98.600000 is a float.
You’re probably surprised that the output doesn’t look like this:
The value %d is an integer.
The value %f is a float.
It doesn’t, because the text included in a printf() function isn’t merely
text — it’s a formatting string.

The printf() function’s formatting string can contain plain text, escape
sequences, and conversion characters, such as the %d in Line 5 and the %f in
Line 6. These conversion characters act as placeholders for values and vari-
ables that follow the formatting string.
For the %d placeholder, the integer value 986 is substituted. The %d conver-
sion character represents integer values.
For the %f placeholder, the float value 98.6 is substituted. The %f conver-
sion character represents floating-point values. Of course, 98.6 isn’t dis-
played. Instead, you see 98.600000. This issue is addressed in the later
section “Minding the extra zeros.”
The %d and %f are only two of many placeholders for the printf() func-
tion’s formatting string. The rest are covered in Chapter 7.