Listing 5-2: The Computer Does the Math
Doing simple arithmetic
Math in your C source code is brought to you by the +, –, *, and / operators.
These are the basic math symbols, with the exception of * and /, mostly
because the × and ÷ characters aren’t found on the typical computer key-
board. For reference, Table 5-1 lists the basic C language math operators.

More C math operators exist, as well as a tumult of mathematical functions.
Chapter 11 helps you continue exploring math programming potential. For
now, the basics will do.

Calculations in C are made by placing values on either side of a math opera-
tor, just as you did all throughout school, but with the benefit of the com-
puter making the calculations. Listing 5-2 provides a sample.