Listing 5-5: Preparing for Constant Updates
Putting constants to use
Anytime your code uses a single value over and over (something significant,
like the number of rows in a table or the maximum number of items you can
stick in a shopping cart), define the value as a constant. Use the #define
directive, as described in the preceding section.
Listing 5-5 shows an update to the source code in Exercise 5-11. The VALUE
constant is created, defined as being equal to 3. Then that constant is used
in the text. The constant is traditionally written in all caps, and you can see
in the source code how doing so makes it easy to find, and identify as, a
constant.