Listing 6-1: Your First Variable
Setting up a quick example
Who likes to read a lot about something before they try it? Not me!
Exercise 6-1: Start a new project in Code::Blocks, name it ex0601, and type
the source code, as shown in Listing 6-1. This project uses a single variable,
x, which is one of the first computer variable names mentioned in the Bible.

Here’s a brief description using the line numbers assigned in the
Code::Blocks editor on the computer screen:
Line 5 contains the variable’s declaration. Every variable used in C must be
declared as a specific variable type and assigned a name. The variable in
Listing 6-1 is declared as an integer (int) and given the name x.
Line 7 assigns the value 5 to variable x. The value goes on the right side of
the equal sign. The variable goes on the left.
Line 8 uses the variable’s value in the printf() statement. The %d conver-
sion character is used because the variable contains an integer value. The
conversion character must match the variable; in Line 8, you see one of each.
Build and run the code. The output looks like this:
The value of variable x is 5.
The following sections describe in further detail the mechanics of creating
and using variables.