Listing 6-2: Working with Variables
Here, variable is the variable’s name. It must be declared earlier in the
source code. value is either an immediate value, a constant, an equation,
another variable, or a value returned from a function. After the statement is
executed, the variable holds the value that’s specified.
Assigning a value to a variable satisfies the second step in using a variable,
but you really need to do something with the variable to make it useful.
Variables can be used anywhere in your source code that a value could oth-
erwise be specified directly.
In Listing 6-2, four variable types are declared, assigned values, and used in
printf() statements.

Type the source code for Listing 6-2 into the editor. Build
and run.
The output looks something like this:
a
1
19.000000
20000.009000
In Line 10, the single character value a is placed into char variable a. Single
characters are expressed using single quotes in C.
In Line 15, you see the %c placeholder used in the printf() statement. That
placeholder is designed for single characters.