Listing 6-3: Oh, No — an Unsigned int!
Using more-specific variable types
The C language’s variable types are more specific than what’s shown in Table
6-1. Depending on the information stored, you may want to use one of these
more detailed variable declarations. Table 6-2 lists a buffet of C language vari-
able types and also the range of values those types can store.

The value range specifies the size of the number you can store in a variable
as well as whether negative numbers are allowed. The compiler may not
always flag warnings that happen when you assign the wrong value to a vari-
able type. So get it right when you declare the variable!
For example, if you need to store the value -10, you use a short int, int,
or long int variable. You cannot use an unsigned int, as the source
code in Listing 6-3 demonstrates.

Here’s the output:
The value of ono is 4294967286.
The moral of the story: If your integer variable stores negative numbers, you
can’t use an unsigned variable type.