Listing 6-4: The Answer to Exercise 6-8
Creating multiple variables
I can find nothing in the rules to prevent starting a section with an exercise,
so here you go:
Exercise 6-8: Create a program that uses the three integer variables shadrach,
meshach, and abednego. Assign integer values to each one, and display the
result.

Here’s a copy of the output from the program generated by Exercise 6-8. It’s
my version of the project:
Shadrach is 701
Meshach is 709
Abednego is 719
Your code can generate different text, but the underlying project should
work. And give yourself a bonus if your answer matched my answer, which is
given in Listing 6-4.

When declaring multiple variables of the same type, you can specify all of
them on the same line, as shown in Listing 6-4 (on Line 5). You don’t even
have to put spaces after each name; the line could have easily been written
int shadrach,meshach,abednego;
The C compiler doesn’t care about spaces — specifically, white space —
outside of something enclosed in double quotes.
I also stacked up the results in a single, long printf() statement. The line
wraps in Listing 6-4 because of this book’s page width, and it may wrap on
your computer screen as well. But if you type code that wraps, don’t press
the Enter key to start a new line.
You can split a long statement in C simply by escaping the Enter key press
at the end of a line. Escaping in this context doesn’t mean that you’re fleeing
danger (other than offending the compiler); instead, you use the backslash
(the escape character) to type the Enter key without messing up your code.
To wit:
printf(“Shad is %d\nMesh is %d\nAbed is d\n”,\
shadrach,meshach,abednego);

I shortened the names in this example so that the text fits on a line on this
page. Between printf()’s formatting string and the variable list, right after
the first comma, I typed a backslash and then pressed the Enter key. The
effect is that the line is broken visually, but the compiler still sees it as a
single statement. Visually, it looks better