Listing 6-6: Variables Recycled
Reusing variables
Variables vary, so their contents can be changed at any time in the program.
The examples shown elsewhere in this chapter use a variable only once and don’t alter its value. That’s pretty much the same as a constant, which makes
them good examples for learning but poor examples for reality.
In your programming journey, variables are declared, and then their values
may be, well, whatever. Not only that; it’s possible to reuse variables over
and over — no harm done. That’s an easy example to show, as illustrated in
Listing 6-6.