Listing 6-7: Variables Mix It Up
Listing 6-7 illustrates how variables can interact with each other.

Line 9 is the one to notice: The value of variable c is assigned the sum of vari-
ables a and b. This calculation is made when the program runs, and then the
result — whatever weirdo value that could be — is displayed.