Listing 7-1: It Eats Characters
Fetching characters with getchar()
It’s time for your code to become more interactive. Consider the source code
from Listing 7-1, which uses the getchar() function. This function gets
(reads) a character from standard input.

The code in Listing 7-1 reads a character from standard input by using the
getchar() function at Line 8. The character is returned from getchar()
and stored in the c integer variable.
Line 9 displays the character stored in c. The printf() function uses the %c
placeholder to display single characters.
Exercise 7-1: Type the source code for project ex0701 , as shown in Listing
7-1. Build and run

The getchar() function is defined this way:
#include <stdio.h>
int getchar(void);
The function has no arguments, so the parentheses are always empty; the
word void in this example basically says so. And the getchar() function
requires the stdio.h header file to be included with your source code.
getchar() returns an integer value, not a char variable. The compiler warns
you when you forget. And don’t worry: The int contains a character value,
which is just how it works.