Listing 7-2: Putting putchar() to Work
Using the putchar() function
The evil twin of the getchar() function is the putchar() function. It serves
the purpose of kicking a single character out to standard output. Here’s the
format:
#include <stdio.h>
int putchar(int c);
To make putchar() work, you send it a character, placing a literal character
in the parentheses, as in
putchar(‘v’);
Or you can place the ASCII code value (an integer) for the character in the
parentheses. The function returns the value of the character that’s written.
(See Listing 7-2.)

This chunk of code uses the getchar() function to pause the program. The
statement in Line 8 waits for input. The input that’s received isn’t stored;
it doesn’t need to be. The compiler doesn’t complain if you don’t keep the
value returned from the getchar() function (or from any function).
In Lines 9 through 12, the putchar() function displays the value of variable
ch one character at a time.
Single-character values are assigned to the ch variable in Lines 9 and 11. The
assignment works just like assigning values, though single characters are
specified, enclosed in single quotes. This process still works, even though ch
isn’t a char variable type.
In Line 13, putchar() displays a constant value directly. Again, the charac-
ter must be enclosed in single quotes.