Listing 7-3: Character Variable Madness
Working with character variables
The getchar() and putchar() functions work with integers, but that
doesn’t mean you need to shun the character variable. The char is still a
variable type in C. When you work with characters, you use the char vari-
able type to store them, as shown in Listing 7-3.

The code declares four char variables at Line 5. These variables are assigned
values in Lines 7 through 10. Line 7 is pretty straightforward. Line 8 uses
math to set the value of variable b to a specific character, as does Line 9 for
variable c. (Use Appendix A to look up a character’s ASCII code value.) Line
10 uses an escape sequence to set a character’s value, something you can’t
type at the keyboard.
All those %c placeholders are stuffed into the printf() statement, but the
output is, well, surprising.