Listing 7-4: Stuffing a String into a char Array
Storing strings

When a program needs text input, it’s necessary to create a place to store
that text. Right away, you’ll probably say, “Golly! That would be a string vari-
able.” If you answered that way, I admire your thinking. You’re relying upon
your knowledge that text in C programming is referred to as a string.
Alas, you’re wrong.
C lacks a string variable type. It does, however, have character variables.
Queue up enough of them and you have a string. Or, to put it in programming
lingo, you have an array of character variables.
The array is a big topic, covered in Chapter 12. For now, be open-minded
about arrays and strings and soak in the goodness of Listing 7-4.

Line 5 creates an array of char variables. The array is a gizmo that lists a
bunch of variables all in a row. The char array variable is named prompt,
which is immediately followed by empty square brackets. It’s the Big Clue
that the variable is an array. The array is assigned, via the equal sign, the text
enclosed in double quotes.
The printf() statement in Line 7 displays the string stored in the prompt
array. The %s conversion character represents the string.
In Line 8, getchar() pauses the program, anticipating the Enter key press.
The program doesn’t follow through by exploding anything, a task I leave up
to you to code at a future date.