Listing 7-5: scanf() Swallows a String
Reading a string with scanf()

One of the most common ways to put the scanf() function to use is to read
in a chunk of text from standard input. To meet that end, the %s conversion
character is used — just like in printf(), but with input instead of output.
(See Listing 7-5.)

Line 5 declares a char array — a string variable — named firstname. The
number in the brackets indicates the size of the array, or the total number of
characters that can be stored there. The array isn’t assigned a value, so it’s
created empty. Basically, the statement at Line 5 sets aside storage for up to
15 characters.

The scanf() function in Line 8 reads a string from standard input and
stores it in the firstname array. The %s conversion character directs
scanf() to look for a string as input, just as %s is a placeholder for strings in
printf()’s output.