Listing 7-6: scanf() Eats an Integer
Reading values with scanf()
The scanf() function can do more than read strings. It can read in any value
specified by a conversion character, as demonstrated in Listing 7-6.

In Listing 7-6, the scanf() function reads in an integer value. The %d conver-
sion character is used, just like printf() — indeed, it’s used in Line 9. That
character directs scanf() to look for an int value for variable fav.

Perhaps you’re wondering about the ampersand (&) in the scanf() function.
The character is a C operator — specifically, the memory address operator.
It’s one of the advanced features in C that’s related to pointers. I avoid the
topic of pointers until Chapter 18, but for now, know that an ampersand must
prefix any variable specified in the scanf() function. The exception is an
array, such as the firstname char array in Listing 7-5.
Try running the program again, but specify a decimal value, such as 41.9, or
type text instead of a number.
The reason you see incorrect output is that scanf() is very specific. It
fetches only the variable type specified by the conversion character. So if
you want a floating-point value, you must specify a float variable and use
the appropriate conversion character; %f, in that case.