Listing 7-7: The fgets() Function Reads a String
Using fgets() for text input
For a general-purpose text input function, one that reads beyond the first
white space character, I recommend the fgets() function. Here’s the
format:
#include <stdio.h>
char * fgets(char *restrict s, int n, FILE *restrict
stream);
Frightening, no? That’s because fgets() is a file function, which reads
text from a file, as in “file get string.” That’s how programmers talk after an
all-nighter.
File functions are covered in Chapter 22, but because the operating system
considers standard input like a file, you can use fgets() to read text from
the keyboard.

Here’s a simplified version of the fgets() function as it applies to reading
text input:
fgets(string,size,stdin);
In this example, string is the name of a char array, a string variable; size
is the amount of text to input plus one, which should be the same size as the
char array; and stdin is the name of the standard input device, as defined
in the stdio.h header file. (See Listing 7-7.)

The fgets() function in Line 8 reads in text. The text goes into the name
array, which is set to a maximum of ten characters in Line 5. The number
10 specifies that fgets() reads in only nine characters, one less than the
number specified. Finally, stdin is specified as the “file” from which input is
read. stdin is standard input.
The char array must have one extra character reserved for the \0 at the end
of a string. Its size must equal the size of input you need — plus one.
Here’s how the program runs on my screen:
Who are you? Andre
Glad to meet you, Andre.
Only the first nine characters of the text I typed in the first line are displayed.
Why only nine? Because of the string’s terminating character — the NULL, or
\0. The room for this character is defined when the name array is created in
Line 5. If fgets() were to read in ten characters instead of nine, the array
would overflow, and the program could malfunction.