Listing 8-1: A Simple Comparison
If What?

All human drama is based on disobedience. No matter what the rules, no
matter how strict the guidelines, some joker breaks free and the rest is an interesting story. That adventure begins with the simple human concept of “what if.” It’s the same concept used for decision making in your programs, though in that instance only the word if is required.

Making a simple comparison

You make comparisons all the time. What will you wear in the morning?
Should you avoid Bill’s office because the receptionist says he’s “testy”
today? And how much longer will you put off going to the dentist? The
computer is no different, albeit the comparisons it makes use values, not
abstracts. (See Listing 8-1.)

Create a new project using the source code shown in Listing 8-1.
Build and run. Here’s the output you should see:
6 is greater than 4
Fast and smart, that’s what a computer is. Here’s how the code works:
Line 5 declares two integer variables: a and b. The variables are assigned
values in Lines 7 and 8, with the value of variable b being calculated to be 2
less than variable a.
Line 10 makes a comparison:
if( a > b)
Programmers read this line as, “If a is greater than b.” Or when they’re teach-
ing the C language, they say, “If variable a is greater than variable b.” And, no,
they don’t read the parentheses.
Lines 11 through 13 belong to the if statement. The meat in the sandwich is
Line 12; the braces (curly brackets) don’t play a decision-making role, other
than hugging the statement at Line 12. If the comparison in Line 10 is true,
the statement in Line 12 is executed. Otherwise, all statements between the
braces are skipped.