Listing 8-2: Values Are Compared
Introducing the if keyword
The if keyword is used to make decisions in your code based upon simple
comparisons. Here’s the basic format:
if(evaluation)
{
statement;
}
The evaluation is a comparison, a mathematical operation, the result of a
function or some other condition. If the condition is true, the statements
(or statement) enclosed in braces are executed; otherwise, they’re skipped.

Comparing values in various ways
The C language employs a small platoon of mathematical comparison opera-
tors. I’ve gathered the bunch in Table 8-1 for your perusal.

Comparisons in C work from left to right, so you read a >= b as “a is greater
than or equal to b.” Also, the order is important: Both >= and <= must be
written in that order, as must the != (not equal) operator. The == operator
can be written either way. (See Listing 8-2.)

The most common comparison is probably the double equal sign. It may look
odd to you. The == operator isn’t the same as the = operator. The = operator
is the assignment operator, which sets values. The == operator is the compari-
son operator, which checks to see whether two values are equal. (See Listing
8-3.)
I pronounce == as “is equal to.”