Listing 8-3: Get “Is Equal To” into Your Head
The most common comparison is probably the double equal sign. It may look
odd to you. The == operator isn’t the same as the = operator. The = operator
is the assignment operator, which sets values. The == operator is the compari-
son operator, which checks to see whether two values are equal. (See Listing
8-3.)
I pronounce == as “is equal to.”