Listing 8-4: Always True
The output may puzzle you. What I see is this:
-3 equals -3
That’s true, isn’t it? But what happened?
Simple: In Line 9, variable a is assigned the value -3. Because that statement
is inside the parentheses, it’s evaluated first. The result of a variable assign-
ment in C is always true for any non-zero value.