Listing 8-5: Semicolon Boo-Boo
Forgetting where to put the semicolon
Listing 8-5 is based upon Listing 8-4, taking advantage of the fact that C
doesn’t require a single statement belonging to an if comparison to be
lodged between curly brackets.

Here’s the output I see:
5 equals -3
The problem here is a common one, a mistake made by just about every C
programmer from time to time: The trailing semicolon in Listing 8-5 (Line 10)
tells the program that the if statement has nothing to do when the condition
is true. That’s because a single semicolon is a complete statement in C, albeit
a null statement. To wit:
if(condition)
;
This construct is basically the same as Line 10 in Listing 8-5. Be careful not
to make the same mistake — especially when you type code a lot and you’re
used to ending a line with a semicolon.