Listing 8-6: An if-else Comparison
Multiple Decisions

Not every decision is a clean-cut, yes-or-no proposition. Exceptions happen
all the time. C provides a few ways to deal with those exceptions, allowing
you to craft code that executes based on multiple possibilities.

Making more-complex decisions

For the either-or type of comparisons, the if keyword has a companion —
else. Together, they work like this:

if(condition)
{
statement(s);
}
else
{
statement(s);
}

if(a==b);
printf(“%d equals %d\n”,a,b);
return(0);

}