Listing 8-7: Logic Is a Tweeting Bird
Building a logical comparison
It’s possible to load two or more comparisons into a single if statement. The
results of the comparisons are then compared by using a logical operator.
When the result of the entire thing is true, the if condition is considered
true. (See Listing 8-7.)

Two comparisons are made by the if statement condition in Line 9. That
statement reads like this: “If the value of variable coordinate is greater than or
equal to –5 and less than or equal to 5.