Listing 8-8: Multiple Choice
Making a multiple-choice selection
The switch-case structure allows you to code decisions in a C program based
on a single value. It’s the multiple-choice selection statement, as shown in
Listing 8-8.

Examine the source code in your editor, where you can reference the line
numbers mentioned in the following paragraphs.
The switch-case structure starts at Line 10 with the switch statement.
The item it evaluates is enclosed in parentheses. Unlike an if statement,
switch eats only a single value. In Line 10, it’s an integer that the user types
(read in Line 8).
The case part of the structure is enclosed in curly brackets, between Lines
11 and 23. A case statement shows a single value, such as 1 in Line 12. The
value is followed by a colon.

The value specified by each case statement is compared with the item speci-
fied in the switch statement. If the values are equal, the statements belong-
ing to case are executed. If not, they’re skipped and the next case value is
compared.
The break keyword stops program flow through the switch-case struc-
ture. Program flow resumes after the switch-case structure’s final curly
bracket, which is Line 24 in Listing 8-8.
After the final comparison, the switch-case structure uses a default item,
shown in Line 21. That item’s statements are executed when none of the
case comparisons matches. The default item is required in the switch-case
structure.

Understanding the switch-case structure
And now — presenting the most complex thing in C. Seriously, you’ll find
more rules and structure with switch-case than just about any other con-
struct in C. Here’s the skeleton:
switch(expression)
{
case value1:
statement(s);
break;
case value2:
statement(s);
break;
case value3:
statement(s);
break;
default:
statement(s);
}

The switch item introduces the structure, which is enclosed by a pair of
curly brackets. The structure must contain at least one case statement and
the default statement.
The switch statement contains an expression in parentheses. That expres-
sion must evaluate to a single value. It can be a variable, a value returned
from a function, or a mathematical operation.
A case statement is followed by an immediate value and then a colon.
Following the colon are one or more statements. These statements are exe-
cuted when the immediate value following case matches the switch state-
ment’s expression. Otherwise, the statements are skipped, and the next case
statement is evaluated.
The break keyword is used to flee the switch-case structure. Otherwise,
program execution falls through the structure.
The default item ends the switch-case structure. It contains statements
that are executed when none of the case statements matches. Or, when
nothing is left to do, the default item doesn’t require any statements — but
it must be specified.
The case portion of a switch-case structure doesn’t make an evaluation. If
you need multiple comparisons, use a multiple if-else type of structure.