Listing 8-10: And Then It Gets Weird
The Weird ?: Decision Thing

I have one last oddball decision-making tool to throw at you in this long, deci-
sive chapter. It’s perhaps the most cryptic of the decision-making tools in C, a
favorite of programmers who enjoy obfuscating their code. Witness Listing 8-10.

Specifically, you want to look at Line 12, which I’m showing here as though it
isn’t ugly enough inside Listing 8-10:
larger = (a > b) ? a : b;

Officially, ?: is known as a ternary operator: It’s composed of three parts. It’s
a comparison and then two parts: value-if-true and value-if-false. Written in
plain, hacker English, the statement looks like this:
result = comparison ? if_true : if_false;
The statement begins with a comparison. Any comparison from an if state-
ment works, as do all operators, mathematical and logical. I typically enclose
the comparison in parentheses, though that’s not a requirement.
When comparison is true, the if_true portion of the statement is evalu-
ated and that value stored in the result variable. Otherwise, the if_false
solution is stored.