Listing 9-1: Write That Down Ten Times!
A Little Déjà Vu
A loop is a section of code that repeats. How often? That depends on how you
write the loop. Basically speaking, a loop involves three things:
✓ Initialization
✓ One or more statements that repeat
✓ An exit
The initialization sets up the loop, usually specifying a condition upon which
the loop begins or is activated. For example, “Start the counter at 1.”
The statements that repeat are contained in curly brackets. They continue to
be executed, one after the other, until the exit condition is met.

The exit condition determines when the loop stops. Either it’s a condition
that’s met, such as “Stop when the counter equals 10,” or the loop can stop
when a break statement is encountered. The program execution continues
with the next statement after the loop’s final curly bracket.
Having an exit condition is perhaps the most important part of a loop. Without
it, the loop repeats forever in a condition called an endless loop. See the later
section “Looping endlessly.”
The C language features two looping keywords: for and while. Assisting the
while keyword is the do keyword. The goto keyword can also be used for
looping, though it’s heavily shunned.

The Thrill of for Loops
A loop is simply a group of statements that repeats. You may want a set
number of iterations, or the number of repeats can be based on a value.
Either way, the for keyword helps set up that basic type of loop.
Doing something x number of times
It’s entirely possible, and even a valid solution, to write source code that dis-
plays the same line of text ten times. You could copy and paste a printf()
statement multiple times to do the job. Simple, but it’s not a loop. (See Listing
9-1.)

As output, the program coughs up the phrase Sore shoulder surgery ten times,
in ten lines of text. The key, of course, is in Line 7, the for statement. That
statement directs the program to repeat the statement(s) in curly brackets a
total of ten times.

Introducing the for loop
The for loop is usually the first type of loop you encounter when you
learn to program. It looks complex, but that’s because it’s doing everything
required of a loop — in a single statement:
for(initialization; exit_condition; repeat_each)
Here’s how it works:
initialization is a C language statement that’s evaluated at the start of
the loop. Most often, it’s where the variable that’s used to count the loop’s
iterations is initialized.
exit_condition is the test upon which the loop stops. In a for loop,
the statements continue to repeat as long as the exit condition is true. The
expression used for the exit_condition is most often a comparison, similar to
something you’d find in an if statement.
repeat_each is a statement that’s executed once every iteration. It’s nor-
mally an operation affecting the variable that’s initialized in the first part of
the for statement.
The for statement is followed by a group of statements held in curly brackets:
for(x=0; x<10; x=x+1)
{
puts(“Sore shoulder surgery”);
}
You can omit the brackets when only one statement is specified:
for(x=0; x<10; x=x+1)
puts(“Sore shoulder surgery”);
In this for statement, and from Listing 9-1, the first expression is initialization:
x=0

The value of the int variable x is set to 0. In C programming, you start count-
ing with 0, not with 1. You’ll see the advantages of doing so as you work
through this book.
The second expression sets the loop’s exit condition:
x<10
As long as the value of variable x is less than 10, the loop repeats. Once that
condition is false, the loop stops. The end effect is that the loop repeats ten
times. That’s because x starts at 0, not at 1.
Finally, here’s the third expression:
x=x+1
Every time the loop spins, the value of variable x is increased by 1. The pre-
ceding statement reads, “Variable x equals the value of variable x, plus 1.”
Because C evaluates the right side of the equation first, nothing is goofed up.
So if the value of x is 5, the code is evaluated as
x=5+1
The new value of x would be 6.
All told, I read the expression this way:
for(x=0; x<10; x=x+1)
“For x starts at 0, while x is less than 10, increment x.”