Listing 9-6: The while Version of Listing 9-1
The Joy of the while Loop
Another popular looping keyword in C is while. It has a companion, do, so
programmers refer to this type of loop as either while or do-while. The C
language is missing the do-whacka-do type of loop.
Structuring a while loop
The C language while loop is a lot easier to look at than a for loop, but it
involves more careful setup and preparation. Basically, it goes like this:
while(condition)
{
statement(s);
}
The condition is a true/false comparison, just like you’d find in an if state-
ment. The condition is checked every time the loop repeats. As long as
it’s true, the loop spins and the statement (or statements) between the curly
brackets continues to execute.
Because the evaluation (condition) happens at the start of the loop, the loop
must be initialized before the while statement, as shown in Listing 9-6.
So how does a while loop end? The termination happens within the loop’s
statements. Usually, one of the statements affects the evaluation, causing it
to turn false.

After the while loop is done, program execution continues with the next
statement after the final curly bracket.
A while loop can also forgo the curly brackets when it has only one statement:
while(condition)
statement;

The while loop demonstrated in Listing 9-6 has three parts:
✓ The initialization takes place on Line 7, where variable x is set equal to 0.
✓ The loop’s exit condition is contained within the while statement’s
parentheses, as shown in Line 8.
✓ The item that iterates the loop is found on Line 11, where variable x
is increased in value. Or, as programmers would say, “Variable x is
incremented.”