Listing 9-7: A Fibonacci Sequence
Using the do-while loop
The do-while loop can be described as an upside-down while loop. That’s
true, especially when you look at the thing’s structure:
do
{
statement(s);
} while (condition);
As with a while loop, the initialization must take place before entering the
loop, and one of the loop’s statements should affect the condition so that
the loop exits. The while statement, however, appears after the last curly
bracket. The do statement begins the structure.
Because of its inverse structure, the major difference between a while loop
and a do-while loop is that the do-while loop is always executed at least
one time. So you can best employ this type of loop when you need to ensure
that the statements spin once. Likewise, avoid do-while when you don’t
want the statements to iterate unless the condition is true. (See Listing 9-7.)

Here’s the output:
0 1 1 2 3 5 8 13 21 34 55 89 144 233
The loop begins at Lines 7 and 8, where the variables are initialized.
Lines 12 through 15 calculate the Fibonacci values. Two printf() functions
display the values.
The loop ends on Line 16, where the while statement makes its evaluation.
As long as variable nacci is less than 300, the loop repeats. You can adjust
this value higher to direct the program to output more Fibonacci numbers.
On Line 18, the putchar() statement cleans up the output by adding a new-
line character.