Listing 9-8: A Common Way to Make an Endless Loop
Loopy Stuff
I could go on and on about loops all day, repeating myself endlessly! Before
moving on, however, I’d like to go over a few looping tips and pratfalls. These
things you should know before you get your official For Dummies Looping
Programmer certificate.
Looping endlessly
Beware the endless loop!
When a program enters an endless loop, it either spews output over and over
without end or it sits there tight and does nothing. Well, it’s doing what you
ordered it to do, which is to sit and spin forever. Sometimes, this setup is done
on purpose, but mostly it happens because of programmer error. And with the
way loops are set up in C, it’s easy to unintentionally loop ad infinitum.
Listing 9-8 illustrates a common endless loop, which is a programming error,
not a syntax error.

The problem with the code in Listing 9-8 is that the for statement’s exit con-
dition is always true: x=10. Read it again if you didn’t catch it the first time,
or just do Exercise 9-18.

The compiler may warn you about the constant TRUE condition in the for
statement. Code::Blocks should do that, and any other compiler would, if
you ratcheted up its error-checking. Otherwise, the program compiles and
runs — infinitely.
✓ To break out of an endless loop, press Ctrl+C on the keyboard. This
trick works only for console programs, and it may not always work. If
it doesn’t, you need to kill the process run amok, which is something I
don’t have time to explain in this book.
✓ Endless loops are also referred to as infinite loops.