Listing 9-10: A for Loop with No Body
Screwing up a loop
I know of two common ways to mess up a loop. These trouble spots crop up
for beginners and pros alike. The only way to avoid these spots is to keep a
keen eye so that you can spot ’em quick.
The first goof-up is specifying a condition that can never be met; for example:
for(x=1;x==10;x=x+1)
In the preceding line, the exit condition is false before the loop spins once,
so the loop is never executed. This error is almost as insidious as using an
assignment operator (a single equal sign) instead of the “is equal to” operator
(as just shown).
Another common mistake is misplacing the semicolon, as in
for(x=1;x<14;x=x+1);
{
puts(“Sore shoulder surgery”);
}
Because the first line, the for statement, ends in a semicolon, the compiler
believes that the line is the entire loop. The empty code repeats 13 times,
which is what the for statement dictates. The puts() statement is then
executed once.
Those rogue semicolons can be frustrating!
The problem is worse with while loops because the do-while structure
requires a semicolon after the final while statement. In fact, forgetting that
particular semicolon is also a source of woe. For a traditional while loop,
you don’t do this:
while(x<14);
{
puts(“Sore shoulder surgery”);
}
The severely stupid thing about these semicolon errors is that the compiler
doesn’t catch them. It believes that your intent is to have a loop without
statements. Such a thing is possible, as shown in Listing 9-10.

In the example shown in Listing 9-10, the semicolon is placed on the line after
the for statement (refer to Line 8 in Listing 9-10). That shows deliberate intent.
You can see that two items are placed in the for statement’s parentheses,
both separated by a comma. That’s perfectly legal, and it works, though it’s
not quite readable.

Though you can load up items in a for statement’s parentheses, it’s rare and
definitely not recommended, for readability’s sake.